requirements.txt
numpy
faiss-cpu
sentence-transformers
requests
accelerate
nltk
rouge-score
bert-score
pdfplumber
bitsandbytes